backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
mongoengine==0.28.2
python-dotenv==1.0.0
Werkzeug==3.0.1
groq==0.4.1
python-docx==1.1.0
reportlab==4.0.9
pymongo==4.6.1
email-validator==2.1.0